refactoring: move erddap and sql model to interface folder, move jsonp config to config folder.  Updated models to current DB representaion for UI.  Removed View folder, full stack testing needs to be done with ooi-ui.
