+++ ooiservices/model/platform_deployment.py
@@ -6,9 +6,9 @@ PlatformDeploymentModel
 '''
 __author__ = 'Matt Campbell'
 
-from ooiservices.model.sqlmodel import SqlModel
+from ooiservices.model.interface.sqlmodel import SqlModel
 
 class PlatformDeploymentModel(SqlModel):
 
     def __init__(self):
-        SqlModel.__init__(self, table_name='platform_deployments', where_param='id')
+        SqlModel.__init__(self, table_name='platform_deployments', where_param='array_code')
