+++ ooiservices/model/stream.py
@@ -4,17 +4,11 @@ ooiservices.model.stream.py
 
 StreamModel
 '''
-
-from ooiservices import app
-from ooiservices.model.sqlmodel import SqlModel
-
 __author__ = "Matt Campbell"
+
+from ooiservices.model.interface.sqlmodel import SqlModel
 
 class StreamModel(SqlModel):
 
     def __init__(self):
-        SqlModel.__init__(self, table_name='streams')
-
-    class Parameter(SqlModel):
-        def __init__(self):
-            SqlModel.__init__(self, table_name='stream_parameters')
+        SqlModel.__init__(self, table_name='streams', where_param='instrument_deployment_code')
