--- ooiservices/view/test.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8">
-    <title>title</title>
-    <script type="text/javascript" src="jquery-2.1.1.min.js"></script>
-    <script type="text/javascript" src="wellknown.js"></script>
-    <script type="text/javascript" src="ooiservices.js"></script>
-  </head>
-  <body>
-    <div id="204">
-        <button onclick="getPlatform('BB00BBBB-BBDDD')">Test 204!</button>
-    </div>
-    <div id="oneTwoOne">
-        <button onclick="getArrays()">Array!</button>
-        <button onclick="getPlatform('GS01SUMO-SM001')">Platform!</button>
-        <button onclick="getInstrument('GS01SUMO-SBD11-01-MOPAK0000')">Instrument!</button>
-    </div>
-    <div id="search">
-       <button onclick="getPlatformsAtArray('GS')">Get platforms at array</button>
-       <button onclick="getInstrumentsAtPlatform('GS03FLMA-RIS02')">Get instruments at platform</button>
-    </div>
-    <div id="out">
-    </div>
-  </body>
-</html>
